need unsigned types
git-svn-id: https://svn.r-project.org/R/trunk@59098 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/Renviron.c
@@ -36,9 +36,10 @@
 /* remove leading and trailing space */
 static char *rmspace(char *s)
 {
-    size_t   i;
+    int i;
 
-    for (i = strlen(s) - 1; i >= 0 && isspace((int)s[i]); i--) s[i] = '\0';
+    for (i = (int) strlen(s) - 1; 
+	 i >= 0 && isspace((int)s[i]); i--) s[i] = '\0';
     for (i = 0; isspace((int)s[i]); i++);
     return s + i;
 }
